+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [Unreleased]
+## [0.7.1] - 2025-06-11
+
+### Added
+- Added Source Detail information for the severity rating output
 
 ### Removed
 - Removed License.md file and integrated it in the README.md due to dual licensing
@@ -19,7 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix minor typo and missing date in Changelog.md
 - Fixed `VEX2PDF_SHOW_OSS_LICENSES` env variable handlers to show updated license information
 
-# [0.7.0] - 2025-05-28
+
+## [0.7.0] - 2025-05-28
 
 ### Added
 - Added `VEX2PDF_REPORT_TITLE` environment variable to override the default report title